build fix attempt 2 (prev sent ios red)
Buddy: Sam

+++ src/NavRouting/SdkModel/NavRoutingPolylineFactory.cpp
@@ -123,17 +123,17 @@ namespace ExampleApp
                     return true;
                 }
 
-                std::vector<std::pair<size_t, size_t>> BuildAmalgamationRanges(const std::vector<NavRoutingPolylineCreateParams>& polylineCreateParams)
+                std::vector<std::pair<int, int>> BuildAmalgamationRanges(const std::vector<NavRoutingPolylineCreateParams>& polylineCreateParams)
                 {
-                    std::vector<std::pair<size_t, size_t>> ranges;
+                    std::vector<std::pair<int, int>> ranges;
 
                     if (polylineCreateParams.empty())
                     {
                         return ranges;
                     }
 
-                    size_t rangeStart = 0;
-                    for (size_t i = 1; i < polylineCreateParams.size(); ++i)
+                    int rangeStart = 0;
+                    for (int i = 1; i < polylineCreateParams.size(); ++i)
                     {
                         const auto &a = polylineCreateParams[i - 1];
                         const auto &b = polylineCreateParams[i];
@@ -144,7 +144,7 @@ namespace ExampleApp
                             rangeStart = i;
                         }
                     }
-                    ranges.push_back(std::make_pair(rangeStart, polylineCreateParams.size()));
+                    ranges.push_back(std::make_pair(rangeStart, int(polylineCreateParams.size())));
 
                     return ranges;
                 }
